Change func name to Upload to match rekor
